[#14] fix : 이벤트 삭제 요청 시 이미 시작된 이벤트면 예외던지도록 수정
커스텀 밸리데이션에 주석 추가

+++ src/main/java/com/sparta/popupstore/domain/promotionevent/service/PromotionEventService.java
@@ -68,7 +68,10 @@ public class PromotionEventService {
 
     @Transactional
     public void deletePromotionEvent(Long promotionEventId) {
-        this.getPromotionEvent(promotionEventId);
+        PromotionEvent promotionEvent = this.getPromotionEvent(promotionEventId);
+        if(promotionEvent.getStartDateTime().isBefore(LocalDateTime.now())){
+            throw new IllegalArgumentException("이미 시작한 이벤트는 삭제할 수 없습니다.");
+        }
         promotionEventRepository.deletePromotionEvent(promotionEventId);
     }
